refactor: rename TransactionTemplate component
